feat: добавить авторизацию

+++ controllers/cards.js
@@ -1,38 +1,48 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-error');
+const ForbiddenError = require('../errors/forbidden-error');
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .populate(['owner', 'likes'])
     .then((cards) => res.send({ data: cards }))
-    .catch(() => res.status(500).send({ message: 'Что-то пошло не так' }));
+    .catch(next);
 };
 
 const deleteCardById = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(new Error())
-    .then(() => res.send({ message: 'Карточка удалена' }))
-    .catch((err) => next(err));
+  Card.findById(req.params.cardId)
+    .orFail(new NotFoundError('Карточка с указанным id не найдена'))
+    .then((card) => {
+      const cardOwner = card.owner.toString();
+      if (cardOwner === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then(() => res.send({ message: 'Карточка удалена' }))
+          .catch(next);
+      }
+      throw new ForbiddenError('Вы не можете удалять карточки других пользователй')
+    })
+    .catch(next);
 };
 
 const createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.status(201).send({ data: card }))
-    .catch((err) => next(err));
+    .catch(next);
 };
 
 const putLikeById = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .orFail(new Error())
+    .orFail(new NotFoundError('Карточка с указанным id не найдена'))
     .then((card) => res.send({ data: card }))
-    .catch((err) => next(err));
+    .catch(next);
 };
 
 const deleteLikeById = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .orFail(new Error())
+    .orFail(new NotFoundError('Карточка с указанным id не найдена'))
     .then((card) => res.send({ data: card }))
-    .catch((err) => next(err));
+    .catch(next);
 };
 
 module.exports = {
